+++ src/main/java/cart/controller/ProductController.java
@@ -1,29 +1,42 @@
 package cart.controller;
 
+import cart.dto.ProductRequest;
 import cart.dto.ProductResponse;
 import cart.service.ProductService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
+import java.net.URI;
 import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
+@RequestMapping("/products")
 public class ProductController {
     private final ProductService productService;
-    
-    @GetMapping
-    public ModelAndView findAll(ModelAndView modelAndView) {
-        modelAndView.setViewName("index");
-        modelAndView.addObject("products", productService.findAll());
-        return modelAndView;
+
+    @PostMapping
+    public ResponseEntity<Void> createProduct(@RequestBody @Valid ProductRequest productRequest) {
+        long id = productService.save(productRequest);
+        return ResponseEntity.created(URI.create("/admin/" + id)).build(); // TODO 상품 상세페이지
     }
-    
-    @GetMapping("/products")
+
+    @GetMapping
     public ResponseEntity<List<ProductResponse>> getProducts() {
         return ResponseEntity.ok(productService.findAll());
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<Void> updateProduct(@PathVariable Long id, @RequestBody ProductRequest productRequest) {
+        productService.update(id, productRequest);
+        return ResponseEntity.noContent().build();
+    }
+    
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteProduct(@PathVariable Long id) {
+        productService.delete(id);
+        return ResponseEntity.noContent().build();
     }
 }
