+++ src/main/java/cart/service/ProductService.java
@@ -20,10 +20,10 @@ public class ProductService {
                 .map(ProductResponse::from)
                 .collect(Collectors.toUnmodifiableList());
     }
-    
-    public void save(final ProductRequest productRequest) {
+
+    public long save(final ProductRequest productRequest) {
         Product product = new Product(null, productRequest.getName(), productRequest.getImageUrl(), productRequest.getPrice());
-        productDao.save(product);
+        return productDao.save(product);
     }
     
     public void update(final Long id, final ProductRequest productRequest) {
