+++ src/test/java/cart/controller/ProductControllerTest.java
@@ -10,6 +10,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.HttpStatus;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.web.servlet.view.InternalResourceViewResolver;
 
 import java.util.List;
 
@@ -21,21 +23,47 @@ import static org.mockito.BDDMockito.given;
 class ProductControllerTest {
     @MockBean
     private ProductService productService;
-
+    
     @BeforeEach
     void setUp() {
-        RestAssuredMockMvc.standaloneSetup(new ProductController(productService));
+        InternalResourceViewResolver viewResolver = new InternalResourceViewResolver();
+        viewResolver.setSuffix(".html");
+        RestAssuredMockMvc.standaloneSetup(
+                MockMvcBuilders.standaloneSetup(new ProductController(productService))
+                        .setViewResolvers(viewResolver)
+        );
     }
 
     @Test
-    void 모든_상품_목록을_가져온다() {
+    void 상품을_조회한다() {
         final ProductResponse firstProductResponse = new ProductResponse(1L, "홍고", "https://ca.slack-edge.com/TFELTJB7V-U04M4NFB5TN-e18b78fabe81-512", 1_000_000_000);
         final ProductResponse secondProductResponse = new ProductResponse(2L, "아벨", "https://ca.slack-edge.com/TFELTJB7V-U04LMNLQ78X-a7ef923d5391-512", 1_000_000_000);
         given(productService.findAll()).willReturn(List.of(firstProductResponse, secondProductResponse));
 
         RestAssuredMockMvc.given().log().all()
-                .when().get("/")
+                .when().get("/admin")
                 .then().log().all()
                 .status(HttpStatus.OK);
+    }
+
+    @Test
+    void 상품을_생성한다() {
+        RestAssuredMockMvc.given().log().all()
+                .when().post("/products")
+                .then().log().all();
+    }
+    
+    @Test
+    void 상품을_수정한다() {
+        RestAssuredMockMvc.given().log().all()
+                .when().put("/products/1")
+                .then().log().all();
+    }
+    
+    @Test
+    void 상품을_삭제한다() {
+        RestAssuredMockMvc.given().log().all()
+                .when().delete("/products/1")
+                .then().log().all();
     }
 }
